Update landing page
Edited the landing page to link people directly to the overview and quickstart (It's not clear when you land on the documentation page that you haven't started at overview). Have saved the specific details on neurodesktop and neurocommand to add to those respective subpages. 

NB. The image links for Acknowledgements aren't working right now.