Update _index.md
Added link to the 2-minutes video, because one video is worth a 1000 words.
I also want to include the image that we have in the "What's next" section (https://neurodesk.github.io/docs/neurodesktop/whats-next/), but I don't know how to embed images in docsy. Can you explain me how to do it?