Update _index.md
Clarified that to upgrade to latest version, users need to first remove the current neurodesktop and then re-install it. I also think that the title in the platform-specific doc pages should be "removing neurodesktop" rather than "stopping neurodesktop", as users need to understand that stuff in the home directory, for example, will be deleted.