Update whats next.md
Clarified that utilities run from the command line may incur a short delay at first use. People might not be prepared for that if they are used for faster response time using the package outside of Neurodesk. If the delay is not related to the CVMFS, please ignore this pull request, and I'll let you know next time I see it. Maybe it's due to some technical issue unrelated to CVMFS.