Update _index.md
I assume that the Nectar Desktop Service does not always have the latest version of Neurodesktop. So I just added a line so people will know that. Or is it the case that every time somebody starts a new Neurodesktop in the Nectar Desktop Service, she/he will get the latest menu?