Update Troubleshooting.md
I made a lot of changes to help users that try to copy and paste to the terminal or the file browser, as they have their unique quirks. The section for the Mac users come next, and if nothing helps, I instruct users how to try out copying and pasting to Mousepad (one of the only "normal" applications, that has both a menu and standard keyboard shortcuts), and report to us.
I hope this will solve most users problems, as the clipboard is really a pain in the ass right now.