Update Troubleshooting.md
This is the last fix (I hope). Sorry for doing it in stages, but things come to my mind on the go.
I changed the titles of the sections a bit, because Clipboard problems might be either within Neurodesktop or between the host computer and Neurodesktop, and each have a different solution.