Update Troubleshooting.md
Removed the example of what is a vertical quotation mark, because even Github gets it wrong (in the markup it's vertical, but in the formatted version it's slightly leaning). Instead, I say "strictly. vertical". If you know how to enforce a real vertical quote in the formatted version, go for it!